feat: implement IP and domain risk calculation functions with enhanced error handling
