+++ src/services/handler/error_interface.ts
@@ -2,8 +2,8 @@ import { CriminalResponseType } from "../../../types/searchIPResponse/CriminalIP
 
 class AppError extends Error {
   constructor(
-    message: string,
     public statusCode: number = 500,
+    message: string,
     public source?: string,
     public details?: any
   ) {
@@ -12,33 +12,39 @@ class AppError extends Error {
     Error.captureStackTrace(this, this.constructor);
   }
 
-  static async fromApiResponse(response: any, source: string): Promise<AppError> {
+  static async fromApiResponse(
+    response: any,
+    source: string
+  ): Promise<AppError> {
     switch (source) {
       case "AbuseIPDB": {
         return new AppError(
-          `AbuseIPDB API Error: ${response.statusText || "Unknown error"}`,
           response.status,
+          `AbuseIPDB API Error: ${response.statusText || "Unknown error"}`,
           source
         );
       }
       case "VirusTotal": {
-          return new AppError(
-            `${source} API Error: ${response.statusText || "Unknown error"}`,
-            response.status,
-            source,
-            response.error
-          );
-        }
-      case "CriminalIP": {
-        const Criminalresponse = (await response.json()) as CriminalResponseType;
         return new AppError(
-          `CriminalIP API Error: ${Criminalresponse.message || "Unknown error"}`,
+          response.status,
+          `${source} API Error: ${response.statusText || "Unknown error"}`,
+          source,
+          response.error
+        );
+      }
+      case "CriminalIP": {
+        const Criminalresponse =
+          (await response.json()) as CriminalResponseType;
+        return new AppError(
           Criminalresponse.status,
+          `CriminalIP API Error: ${
+            Criminalresponse.message || "Unknown error"
+          }`,
           source
         );
       }
       default:
-        return new AppError(`${source} API Error: Unknown error`, 500, source);
+        return new AppError(500, `${source} API Error: Unknown error`, source);
     }
   }
 }
